modified file structure
